Add experimental transaction support
Breaking: @__test_run no longer available in migration context
Breaking: session & migration are now defined methods in a migration context

To use transactions in a per document migration, call use_transaction!

At this stage this is an experimental feature so should be used with
caution. There is some trickiness with migration logs needing to be
updated out of the transaction or the updates will overwrite each other.